Refactor cache interface
