+++ libvmi/Makefile.am
@@ -22,7 +22,6 @@ h_private = \
 
 c_sources = \
     accessors.c \
-    cache.c \
     convenience.c \
     core.c \
     events.c \
@@ -41,6 +40,10 @@ c_sources = \
     driver/driver_interface.c \
     driver/memory_cache.c \
     os/os_interface.c
+
+if ENABLE_ADDRESS_CACHE
+c_sources  += cache.c
+endif
 
 if REKALL_PROFILES
 c_sources  += rekall.c
@@ -108,5 +111,6 @@ lib_LTLIBRARIES= libvmi.la
 libvmi_la_SOURCES= $(h_public) $(h_private) $(drivers) $(os) $(c_sources)
 libvmi_la_LIBADD= config/libconfig.la
 libvmi_la_CFLAGS= -fvisibility=hidden $(GLIB_CFLAGS)
-libvmi_la_CFLAGS+= -Werror -Wall -Wextra -Wno-missing-field-initializers
+libvmi_la_CFLAGS+= -Werror -Wall -Wextra
+libvmi_la_CFLAGS+= -Wno-missing-field-initializers -Wno-unused-but-set-variable
 libvmi_la_LDFLAGS= -release $(RELEASE) $(GLIB_LIBS)
